DjangoRestFramework-Project/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
Markdown==3.3.6
mock==4.0.3
nltk==3.6.7
numpy==1.22.1
pandas==1.4.0
Pillow==9.0.0
pipreqs==0.4.11
psycopg==3.0.8
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.3
regex==2022.1.18
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.8
yarg==0.1.9
zipp==3.7.0
